tests: also split test files
