+++ pkg/frontend/nvme_controller_test.go
@@ -1,0 +1,691 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
+// Copyright (C) 2023 Intel Corporation
+
+// Package frontend implememnts the FrontEnd APIs (host facing) of the storage Server
+package frontend
+
+import (
+	"bytes"
+	"fmt"
+	"net"
+	"reflect"
+	"testing"
+
+	"google.golang.org/protobuf/proto"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/emptypb"
+	"google.golang.org/protobuf/types/known/fieldmaskpb"
+
+	pc "github.com/opiproject/opi-api/common/v1/gen/go"
+	pb "github.com/opiproject/opi-api/storage/v1alpha1/gen/go"
+	"github.com/opiproject/opi-spdk-bridge/pkg/server"
+)
+
+var (
+	testControllerID   = "controller-test"
+	testControllerName = server.ResourceIDToVolumeName(testControllerID)
+	testController     = pb.NvmeController{
+		Spec: &pb.NvmeControllerSpec{
+			SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+			PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+			NvmeControllerId: 17,
+		},
+		Status: &pb.NvmeControllerStatus{
+			Active: true,
+		},
+	}
+)
+
+func TestFrontEnd_CreateNvmeController(t *testing.T) {
+	tests := map[string]struct {
+		id      string
+		in      *pb.NvmeController
+		out     *pb.NvmeController
+		spdk    []string
+		errCode codes.Code
+		errMsg  string
+		start   bool
+		exist   bool
+	}{
+		"illegal resource_id": {
+			"CapitalLettersNotAllowed",
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 1,
+				},
+			},
+			nil,
+			[]string{""},
+			codes.Unknown,
+			fmt.Sprintf("user-settable ID must only contain lowercase, numbers and hyphens (%v)", "got: 'C' in position 0"),
+			false,
+			false,
+		},
+		"valid request with invalid SPDK response": {
+			testControllerID,
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 1,
+				},
+			},
+			nil,
+			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
+			codes.InvalidArgument,
+			fmt.Sprintf("Could not create CTRL: %v", testControllerName),
+			true,
+			false,
+		},
+		"valid request with empty SPDK response": {
+			testControllerID,
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 1,
+				},
+			},
+			nil,
+			[]string{""},
+			codes.Unknown,
+			fmt.Sprintf("nvmf_subsystem_add_listener: %v", "EOF"),
+			true,
+			false,
+		},
+		"valid request with ID mismatch SPDK response": {
+			testControllerID,
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 1,
+				},
+			},
+			nil,
+			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
+			codes.Unknown,
+			fmt.Sprintf("nvmf_subsystem_add_listener: %v", "json response ID mismatch"),
+			true,
+			false,
+		},
+		"valid request with error code from SPDK response": {
+			testControllerID,
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 1,
+				},
+			},
+			nil,
+			[]string{`{"id":%d,"error":{"code":-32602,"message":"Invalid parameters"}}`},
+			codes.Unknown,
+			fmt.Sprintf("nvmf_subsystem_add_listener: %v", "json response error: Invalid parameters"),
+			true,
+			false,
+		},
+		"valid request with valid SPDK response": {
+			testControllerID,
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 17,
+				},
+			},
+			&pb.NvmeController{
+				Name: testControllerName,
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: -1,
+				},
+				Status: &pb.NvmeControllerStatus{
+					Active: true,
+				},
+			},
+			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`},
+			codes.OK,
+			"",
+			true,
+			false,
+		},
+		"already exists": {
+			testControllerID,
+			&pb.NvmeController{
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+					NvmeControllerId: 17,
+				},
+			},
+			&testController,
+			[]string{""},
+			codes.OK,
+			"",
+			false,
+			true,
+		},
+	}
+
+	// run tests
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			testEnv := createTestEnvironment(tt.start, tt.spdk)
+			defer testEnv.Close()
+
+			testEnv.opiSpdkServer.Nvme.Subsystems[testSubsystemName] = &testSubsystem
+			testEnv.opiSpdkServer.Nvme.Namespaces[testNamespaceName] = &testNamespace
+			if tt.exist {
+				testEnv.opiSpdkServer.Nvme.Controllers[testControllerName] = &testController
+			}
+			if tt.out != nil {
+				tt.out.Name = testControllerName
+			}
+
+			request := &pb.CreateNvmeControllerRequest{NvmeController: tt.in, NvmeControllerId: tt.id}
+			response, err := testEnv.client.CreateNvmeController(testEnv.ctx, request)
+			if response != nil {
+				mtt, _ := proto.Marshal(tt.out)
+				mResponse, _ := proto.Marshal(response)
+				if !bytes.Equal(mtt, mResponse) {
+					t.Error("response: expected", tt.out, "received", response)
+				}
+			}
+
+			if err != nil {
+				if er, ok := status.FromError(err); ok {
+					if er.Code() != tt.errCode {
+						t.Error("error code: expected", tt.errCode, "received", er.Code())
+					}
+					if er.Message() != tt.errMsg {
+						t.Error("error message: expected", tt.errMsg, "received", er.Message())
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestFrontEnd_DeleteNvmeController(t *testing.T) {
+	tests := map[string]struct {
+		in      string
+		out     *emptypb.Empty
+		spdk    []string
+		errCode codes.Code
+		errMsg  string
+		start   bool
+		missing bool
+	}{
+		"valid request with invalid SPDK response": {
+			testControllerName,
+			nil,
+			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
+			codes.InvalidArgument,
+			fmt.Sprintf("Could not delete NQN:ID %v", "nqn.2022-09.io.spdk:opi3:17"),
+			true,
+			false,
+		},
+		"valid request with empty SPDK response": {
+			testControllerName,
+			nil,
+			[]string{""},
+			codes.Unknown,
+			fmt.Sprintf("nvmf_subsystem_remove_listener: %v", "EOF"),
+			true,
+			false,
+		},
+		"valid request with ID mismatch SPDK response": {
+			testControllerName,
+			nil,
+			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
+			codes.Unknown,
+			fmt.Sprintf("nvmf_subsystem_remove_listener: %v", "json response ID mismatch"),
+			true,
+			false,
+		},
+		"valid request with error code from SPDK response": {
+			testControllerName,
+			nil,
+			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
+			codes.Unknown,
+			fmt.Sprintf("nvmf_subsystem_remove_listener: %v", "json response error: myopierr"),
+			true,
+			false,
+		},
+		"valid request with valid SPDK response": {
+			testControllerName,
+			&emptypb.Empty{},
+			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`}, // `{"jsonrpc": "2.0", "id": 1, "result": True}`,
+			codes.OK,
+			"",
+			true,
+			false,
+		},
+		"valid request with unknown key": {
+			server.ResourceIDToVolumeName("unknown-controller-id"),
+			nil,
+			[]string{""},
+			codes.NotFound,
+			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-controller-id")),
+			false,
+			false,
+		},
+		"unknown key with missing allowed": {
+			server.ResourceIDToVolumeName("unknown-id"),
+			&emptypb.Empty{},
+			[]string{""},
+			codes.OK,
+			"",
+			false,
+			true,
+		},
+	}
+
+	// run tests
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			testEnv := createTestEnvironment(tt.start, tt.spdk)
+			defer testEnv.Close()
+			testEnv.opiSpdkServer.Nvme.Subsystems[testSubsystemName] = &testSubsystem
+			testEnv.opiSpdkServer.Nvme.Controllers[testControllerName] = &testController
+
+			request := &pb.DeleteNvmeControllerRequest{Name: tt.in, AllowMissing: tt.missing}
+			response, err := testEnv.client.DeleteNvmeController(testEnv.ctx, request)
+			if err != nil {
+				if er, ok := status.FromError(err); ok {
+					if er.Code() != tt.errCode {
+						t.Error("error code: expected", tt.errCode, "received", er.Code())
+					}
+					if er.Message() != tt.errMsg {
+						t.Error("error message: expected", tt.errMsg, "received", er.Message())
+					}
+				}
+			}
+			if reflect.TypeOf(response) != reflect.TypeOf(tt.out) {
+				t.Error("response: expected", reflect.TypeOf(tt.out), "received", reflect.TypeOf(response))
+			}
+		})
+	}
+}
+
+func TestFrontEnd_UpdateNvmeController(t *testing.T) {
+	spec := &pb.NvmeControllerSpec{
+		SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+		PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+		NvmeControllerId: 17,
+	}
+	tests := map[string]struct {
+		mask    *fieldmaskpb.FieldMask
+		in      *pb.NvmeController
+		out     *pb.NvmeController
+		spdk    []string
+		errCode codes.Code
+		errMsg  string
+		start   bool
+	}{
+		"invalid fieldmask": {
+			&fieldmaskpb.FieldMask{Paths: []string{"*", "author"}},
+			&pb.NvmeController{
+				Name: testControllerName,
+				Spec: spec,
+			},
+			nil,
+			[]string{""},
+			codes.Unknown,
+			fmt.Sprintf("invalid field path: %s", "'*' must not be used with other paths"),
+			false,
+		},
+		"valid request without SPDK": {
+			nil,
+			&pb.NvmeController{
+				Name: testControllerName,
+				Spec: spec,
+			},
+			&pb.NvmeController{
+				Name: testControllerName,
+				Spec: spec,
+				Status: &pb.NvmeControllerStatus{
+					Active: true,
+				},
+			},
+			[]string{""},
+			codes.OK,
+			"",
+			false,
+		},
+		"valid request with unknown key": {
+			nil,
+			&pb.NvmeController{
+				Name: server.ResourceIDToVolumeName("unknown-id"),
+				Spec: spec,
+			},
+			nil,
+			[]string{""},
+			codes.NotFound,
+			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-id")),
+			false,
+		},
+	}
+
+	// run tests
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			testEnv := createTestEnvironment(tt.start, tt.spdk)
+			defer testEnv.Close()
+
+			testEnv.opiSpdkServer.Nvme.Controllers[testControllerName] = &testController
+
+			request := &pb.UpdateNvmeControllerRequest{NvmeController: tt.in, UpdateMask: tt.mask}
+			response, err := testEnv.client.UpdateNvmeController(testEnv.ctx, request)
+			if response != nil {
+				// Marshall the request and response, so we can just compare the contained data
+				mtt, _ := proto.Marshal(tt.out.Spec)
+				mResponse, _ := proto.Marshal(response.Spec)
+
+				// Compare the marshalled messages
+				if !bytes.Equal(mtt, mResponse) {
+					t.Error("response: expected", tt.out.GetSpec(), "received", response.GetSpec())
+				}
+				if !reflect.DeepEqual(response.Status, tt.out.Status) {
+					t.Error("response: expected", tt.out.GetStatus(), "received", response.GetStatus())
+				}
+			}
+
+			if err != nil {
+				if er, ok := status.FromError(err); ok {
+					if er.Code() != tt.errCode {
+						t.Error("error code: expected", tt.errCode, "received", er.Code())
+					}
+					if er.Message() != tt.errMsg {
+						t.Error("error message: expected", tt.errMsg, "received", er.Message())
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestFrontEnd_ListNvmeControllers(t *testing.T) {
+	secondSubsystemName := server.ResourceIDToVolumeName("controller-test1")
+	tests := map[string]struct {
+		in      string
+		out     []*pb.NvmeController
+		spdk    []string
+		errCode codes.Code
+		errMsg  string
+		start   bool
+		size    int32
+		token   string
+	}{
+		"valid request without SPDK": {
+			testSubsystemName,
+			[]*pb.NvmeController{
+				{
+					Name: testControllerName,
+					Spec: &pb.NvmeControllerSpec{
+						SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
+						PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
+						NvmeControllerId: 17,
+					},
+					Status: &pb.NvmeControllerStatus{
+						Active: true,
+					},
+				},
+				{
+					Name: secondSubsystemName,
+					Spec: &pb.NvmeControllerSpec{
+						SubsystemId:      &pc.ObjectKey{Value: server.ResourceIDToVolumeName("subsystem-test1")},
+						PcieId:           &pb.PciEndpoint{PhysicalFunction: 2, VirtualFunction: 2},
+						NvmeControllerId: 17,
+					},
+					Status: &pb.NvmeControllerStatus{
+						Active: true,
+					},
+				},
+			},
+			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 1, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 1, "pci_bdf": "ca:00.3"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 2, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 2, "pci_bdf": "ca:00.4"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 3, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 3, "pci_bdf": "ca:00.5"}]}`},
+			codes.OK,
+			"",
+			false,
+			0,
+			"",
+		},
+	}
+
+	// run tests
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			testEnv := createTestEnvironment(tt.start, tt.spdk)
+			defer testEnv.Close()
+			testEnv.opiSpdkServer.Nvme.Subsystems[testSubsystemName] = &testSubsystem
+			testEnv.opiSpdkServer.Nvme.Controllers[testControllerName] = &testController
+			testEnv.opiSpdkServer.Nvme.Controllers[secondSubsystemName] = &pb.NvmeController{
+				Name: secondSubsystemName,
+				Spec: &pb.NvmeControllerSpec{
+					SubsystemId:      &pc.ObjectKey{Value: server.ResourceIDToVolumeName("subsystem-test1")},
+					PcieId:           &pb.PciEndpoint{PhysicalFunction: 2, VirtualFunction: 2},
+					NvmeControllerId: 17,
+				},
+				Status: &pb.NvmeControllerStatus{
+					Active: true,
+				},
+			}
+
+			request := &pb.ListNvmeControllersRequest{Parent: tt.in, PageSize: tt.size, PageToken: tt.token}
+			response, err := testEnv.client.ListNvmeControllers(testEnv.ctx, request)
+			if response != nil {
+				if !reflect.DeepEqual(response.NvmeControllers, tt.out) {
+					t.Error("response: expected", tt.out, "received", response.NvmeControllers)
+				}
+			}
+
+			if err != nil {
+				if er, ok := status.FromError(err); ok {
+					if er.Code() != tt.errCode {
+						t.Error("error code: expected", tt.errCode, "received", er.Code())
+					}
+					if er.Message() != tt.errMsg {
+						t.Error("error message: expected", tt.errMsg, "received", er.Message())
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestFrontEnd_GetNvmeController(t *testing.T) {
+	tests := map[string]struct {
+		in      string
+		out     *pb.NvmeController
+		spdk    []string
+		errCode codes.Code
+		errMsg  string
+		start   bool
+	}{
+		"valid request with valid SPDK response": {
+			testControllerName,
+			&pb.NvmeController{
+				Name: testControllerName,
+				Spec: &pb.NvmeControllerSpec{
+					NvmeControllerId: 17,
+				},
+				Status: &pb.NvmeControllerStatus{
+					Active: true,
+				},
+			},
+			[]string{""},
+			codes.OK,
+			"",
+			false,
+		},
+		"valid request with unknown key": {
+			"unknown-controller-id",
+			nil,
+			[]string{""},
+			codes.NotFound,
+			fmt.Sprintf("unable to find key %s", "unknown-controller-id"),
+			false,
+		},
+	}
+
+	// run tests
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			testEnv := createTestEnvironment(tt.start, tt.spdk)
+			defer testEnv.Close()
+			testEnv.opiSpdkServer.Nvme.Subsystems[testSubsystemName] = &testSubsystem
+			testEnv.opiSpdkServer.Nvme.Controllers[testControllerName] = &testController
+
+			request := &pb.GetNvmeControllerRequest{Name: tt.in}
+			response, err := testEnv.client.GetNvmeController(testEnv.ctx, request)
+			if response != nil {
+				if !reflect.DeepEqual(response.Spec, tt.out.Spec) {
+					t.Error("response: expected", tt.out.GetSpec(), "received", response.GetSpec())
+				}
+				if !reflect.DeepEqual(response.Status, tt.out.Status) {
+					t.Error("response: expected", tt.out.GetStatus(), "received", response.GetStatus())
+				}
+			}
+
+			if err != nil {
+				if er, ok := status.FromError(err); ok {
+					if er.Code() != tt.errCode {
+						t.Error("error code: expected", tt.errCode, "received", er.Code())
+					}
+					if er.Message() != tt.errMsg {
+						t.Error("error message: expected", tt.errMsg, "received", er.Message())
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestFrontEnd_NvmeControllerStats(t *testing.T) {
+	tests := map[string]struct {
+		in      string
+		out     *pb.VolumeStats
+		spdk    []string
+		errCode codes.Code
+		errMsg  string
+		start   bool
+	}{
+		"valid request with valid SPDK response": {
+			testControllerName,
+			&pb.VolumeStats{
+				ReadOpsCount:  -1,
+				WriteOpsCount: -1,
+			},
+			[]string{""},
+			codes.OK,
+			"",
+			false,
+		},
+		"valid request with unknown key": {
+			server.ResourceIDToVolumeName("unknown-id"),
+			nil,
+			[]string{""},
+			codes.NotFound,
+			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-id")),
+			false,
+		},
+	}
+
+	// run tests
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			testEnv := createTestEnvironment(tt.start, tt.spdk)
+			defer testEnv.Close()
+
+			testEnv.opiSpdkServer.Nvme.Controllers[testControllerName] = &testController
+
+			request := &pb.NvmeControllerStatsRequest{Id: &pc.ObjectKey{Value: tt.in}}
+			response, err := testEnv.client.NvmeControllerStats(testEnv.ctx, request)
+			if response != nil {
+				if !reflect.DeepEqual(response.Stats, tt.out) {
+					t.Error("response: expected", tt.out, "received", response.Stats)
+				}
+			}
+
+			if err != nil {
+				if er, ok := status.FromError(err); ok {
+					if er.Code() != tt.errCode {
+						t.Error("error code: expected", tt.errCode, "received", er.Code())
+					}
+					if er.Message() != tt.errMsg {
+						t.Error("error message: expected", tt.errMsg, "received", er.Message())
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestFrontEnd_NewTcpSubsystemListener(t *testing.T) {
+	tests := map[string]struct {
+		listenAddress string
+		wantPanic     bool
+		protocol      string
+	}{
+		"ipv4 valid address": {
+			listenAddress: "10.10.10.10:12345",
+			wantPanic:     false,
+			protocol:      ipv4NvmeTCPProtocol,
+		},
+		"valid ipv6 addresses": {
+			listenAddress: "[2002:0db0:8833:0000:0000:8a8a:0330:7337]:54321",
+			wantPanic:     false,
+			protocol:      ipv6NvmeTCPProtocol,
+		},
+		"empty string as listen address": {
+			listenAddress: "",
+			wantPanic:     true,
+			protocol:      "",
+		},
+		"missing port": {
+			listenAddress: "10.10.10.10",
+			wantPanic:     true,
+			protocol:      "",
+		},
+		"valid port invalid ip": {
+			listenAddress: "wrong:12345",
+			wantPanic:     true,
+			protocol:      "",
+		},
+		"meaningless listen address": {
+			listenAddress: "some string which is not ip address",
+			wantPanic:     true,
+			protocol:      "",
+		},
+	}
+
+	for testName, tt := range tests {
+		t.Run(testName, func(t *testing.T) {
+			defer func() {
+				r := recover()
+				if (r != nil) != tt.wantPanic {
+					t.Errorf("NewTCPSubsystemListener() recover = %v, wantPanic = %v", r, tt.wantPanic)
+				}
+			}()
+
+			gotSubsysListener := NewTCPSubsystemListener(tt.listenAddress)
+			host, port, _ := net.SplitHostPort(tt.listenAddress)
+			wantSubsysListener := &tcpSubsystemListener{
+				listenAddr: net.ParseIP(host),
+				listenPort: port,
+				protocol:   tt.protocol,
+			}
+
+			if !reflect.DeepEqual(gotSubsysListener, wantSubsysListener) {
+				t.Errorf("Expect %v subsystem listener, received %v", wantSubsysListener, gotSubsysListener)
+			}
+		})
+	}
+}
